+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
-## [2.4.1](https://github.com/bitrix24/b24ui/compare/v2.3.0...v2.4.1) (2026-03-04)
+## [2.4.2](https://github.com/bitrix24/b24ui/compare/v2.4.1...v2.4.2) (2026-03-xx)
+
+### Features
+
+### Bug Fixes
+
+* fix(Input|Textarea): padding for `noPadding+loading`
+* fix(components): improve `disabled` state
+
+## [2.4.1](https://github.com/bitrix24/b24ui/compare/v2.4.0...v2.4.1) (2026-03-04)
 
 ### Features